moved cobra files to cmd directory
